Fix ambiguous TestEqual overload in PSPossessionComponentTests
FVector components are double in UE5; comparing against a float ScrimmageX
was ambiguous between the float and double TestEqual overloads. Cast both
sides to double explicitly.

+++ Source/PlaySports/Private/Tests/PSPossessionComponentTests.cpp
@@ -98,7 +98,7 @@ bool FPSC3FieldGridFormationSpawnTest::RunTest(const FString& Parameters)
             else
             {
                 TestEqual(TEXT("Non-QB pawn is spawned on the scrimmage line"),
-                    Pawn->GetActorLocation().X, ScrimmageX);
+                    static_cast<double>(Pawn->GetActorLocation().X), static_cast<double>(ScrimmageX));
             }
         }
     }
